support use of slackbot as a library
